More ES6 stuff and componentize.
